requirements.txt
appdirs==1.4.2
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.5.4
gunicorn==19.7.0
packaging==16.8
pyparsing==2.2.0
whitenoise==3.3.0
six==1.10.0
